Included model for Featured product
I created a model called "FeaturedProduct" which extends "Product" model. I did it to keep record of products that are to be featured. Extending the "Product" model might not be the best solution for this, but it seems to be working fine. It should probably be a foreign key to "Product" model or better yet, not create a new model and instead add this as a field to "Product" model directly as a boolean field...